Cleanup + modal

+++ src/app/core/modal/dialog-wrapper.component.ts
@@ -1,15 +1,51 @@
 import {
-    Component, ViewContainerRef, ViewChild, ComponentFactoryResolver, ReflectiveInjector, Type
+    Component, HostListener, ViewContainerRef, ViewChild, ComponentFactoryResolver, ReflectiveInjector, Type
 } from '@angular/core';
+
 import { DialogComponent } from "./dialog.component";
 import { DialogService } from "./dialog.service";
 
+import { trigger, animate, transition, style, query } from '@angular/animations';
+
+export const fadeAnimation =
+
+    trigger('fadeAnimation', [
+        transition('* => *', [
+            query(':enter',
+                [
+                    style({ opacity: 0 })
+                ],
+                { optional: true }
+            ),
+            query(':leave',
+                [
+                    style({ opacity: 1 }),
+                    animate('0.2s', style({ opacity: 0 }))
+                ],
+                { optional: true }
+            ),
+            query(':enter',
+                [
+                    style({ opacity: 0 }),
+                    animate('0.2s', style({ opacity: 1 }))
+                ],
+                { optional: true }
+            )
+        ])
+    ]);
+
 @Component({
     selector: 'dialog-wrapper',
+    animations: [
+        fadeAnimation
+    ],
     template: `
-    <div #container class="modal" role="dialog">
+    <div #container [@fadeAnimation]="true" class="modal" role="dialog">
         <div class="modal-background"></div>
-        <ng-template #element></ng-template>
+        <div class="modal-content">
+            <ng-template #element></ng-template>
+        </div>
+        <button class="modal-close is-large" (click)="closeDialog()"></button>
     </div>
 `
 })
@@ -30,6 +66,8 @@ export class DialogWrapperComponent {
      * @type {DialogComponent}
      */
     private content: DialogComponent<any, any>;
+
+    private handleEscapePressed: boolean;
 
     /**
      * Constructor
@@ -58,12 +96,25 @@ export class DialogWrapperComponent {
      */
     closeByClickOutside() {
         const containerEl = this.container.nativeElement;
-        containerEl.querySelector('.modal-content').addEventListener('click', (event) => {
+        containerEl.querySelector('.modal-background').addEventListener('click', (event) => {
+            this.dialogService.removeDialog(this.content);
             event.stopPropagation();
         });
-        containerEl.addEventListener('click', () => {
+    }
+
+    closeByEscapeKeyPressed() {
+        this.handleEscapePressed = true;
+    }
+
+    closeDialog() {
+        this.dialogService.removeDialog(this.content);
+    }
+
+    @HostListener('window:keyup', ['$event'])
+    onKeyPress(e: KeyboardEvent) {
+        if(!this.handleEscapePressed)
+            return;
+        if (e.keyCode == 27)
             this.dialogService.removeDialog(this.content);
-        }, false);
     }
 }
-
